requirements.txt
langchain-core
langchain-openai
langchain-community
langchain-text-splitters
pydantic
numpy
langgraph
openai
chromadb
arxiv
pypdf
langchain_groq
streamlit
python-dotenv
langchain_huggingface
langchain_groq
streamlit
nest_asyncio>=1.5.8
